Simplify YAML Config Structure
This moves injectable values so they live closer to the templates that
they affect. This also clarifies that each of the major sections is
actually a template that gets repeated for each permutation, and
simplifies how we specify the array of permutations. I can think of no
reason why the `permutations` key needs to support an array of value
arrays, so I am flattening that to being a single array of values
instead. I am keeping the values under a `values` key in `permutations`,
in case we later need to support dynamic filters for permutations or
some other kind of generator for permutations other than a static list.