Switch `spec` to `mergeSpec` for Volume and Mount Templates
These values are handled differently. Kubernetes does not use a `spec`
key underneath the `volumeMounts` and `volumes` keys of deployments.
Instead, configuration values are intermixed with the standard set of
settings for those settings.

For most other manifest changes described by the KSCT config, to cleanly
separate things, we keep settings that appear verbatim in output
manifests under a separate `spec` key. But, it'd be confusing to call
the key for *volumeMounts* and *volumes* a `spec` since the other `spec`
keys are copied verbatim, while these have to be merged with the larger
spec. So, to make things clearer, I'm calling these `mergeSpec` keys.
That way, hopefully users understand that these values get merged in
with the other values rather than being a standardalone spec.