Rename `volumeMountTemplates` and `mountTemplates`
Volume mounts in Kubernetes are called `volumeMounts`, not `mounts`, so
it doesn't make sense for us to use a different name for those here. At
the same time, it was confusing that both `volumeTemplates` and
`mountTemplates` were under a key called `volumeMountTemplates` when the
settings pertain to more than just `volumeMounts` in the output.

To fix this, we're renaming `volumeMountTemplates` to
`containerVolumeTemplates` since these configurations describe what
changes to make to `volumes` and `volumeMounts` as they relate to
specific containers. That frees up the `volumeMountTemplates` key to be
used instead of `mountTemplates`.