Implement `injectedValues` Handling for PVs
This makes use of the Galbar JSONPath library as a shortcut to parsing
and acting on YAML paths. Since Yaml is a superset of JSON, this should
be fine. It may also allow injected value paths a lot of latitude in how
values in output is populated, since pretty much any JSONPath expression
should work here.

For ease of configuration, we automatically insert the leading `$.` of
the expression. I do not anticipate that being a problem, but we can
change it later if we need.